requirements.txt
streamlit
pandas
openpyxl
python-calamine
fpdf2